Add basic api endpoints for games
A game is a betting game and contains the cup that the game is
about (e.g. European Championship 2016) and how many points the players
get for which events. It also contains the bets that are related to this
game but this is not implemented yet.

Just two GET routes are implemented so far:

- api/games/ --> Show all available games
- api/games/:title --> Show the game with the given :title

The controller methods do not contain any error handling and just work
in a best case scenario (TODO).