Return JSON API conforming data from the api
The JSON API Specification is a widely used (lots of starts on GitHub)
specification for APIs. That's why this API should follow it too.
See: http://jsonapi.org/

To serialize and deserialize the data, a ready made library is used to
avoid reinventing the wheel. Most libraries are bound to a certain
framework, while jsonapi-serializer is framework agnostic.

I saved the serialize options in the model because it ultimately is
closely linked to the model and it makes sense for me to have that
information in one file.

The content negotiation is very minimal for now (does not check if the
client sets the correct headers and so on). Needs to be improved in
later commits.