Enable creating games with a POST to /api/games
To be able to use post requests, the body has to be parsed. The
koa-bodyparser does just that. The JSON API type needs to be added to be
parsed as JSON.

Since the body is going to follow JSON API specs, it needs to be
deserialized. The relationship cup should be saved as a simple databse
field cup and needs a valueForRelationship function to make that
possible.

Some attributes contain multiple words that are in camelCase format. The
deserialize function automatically converts them to dash-case. The
keyForAttribute settings avoids that.

TODO: Secure the POST route to make sure only admins can create new
games.