Secure POST /api/games with basic auth
The creation of resources should only be possible by admins. In a
later stage this should work via a login route. For now a simple basic
auth on the relevant API endpoints is enough.

The special middleware basicAuth makes it possible to use the
koa-basic-auth library with username and password from a .env file (to
avoid commiting credentials to the repository). It is important to use
koa-basic-auth@next to be able to work with koa 2.

The basicAuth middleware can be added as the first middleware of routes
that should be protected.

Since the routes are going to use the basicAuth middleware the
dotenv.config() needs to be called before the app is even loaded (so all
functions of the app have access to the .env variables). That's why it
is moved from app.js to index.js.