Replace Game title with a String _id
The _id field of mongo DB can be changed to be something different than
ObjectId. Since the users should be able to define a unique name for
their game, this name can be used as the _id. This will lead to pretty
urls (e.g. /api/games/my-betting-game).

When a POST to /api/games contains an id field, the deserializer saves
an id attribute. However, mongo needs an _id attribute. That is why the
id attribute is copied into _id. mongoose creates a new Game with the
specified id then (if it does not exist yet).