Embed Bets into Games
Since Bets can't exist without a game and need to be retrieved as part
of the Game most of the time, it is better to embed.
See: https://docs.mongodb.com/manual/tutorial/model-embedded-one-to-many-relationships-between-documents/

The bets api still works as expected, the controller just adds and
retrieves the values from the games collection.

The games serialize method needed adaption to make sure that the bets
are exported and their content is included in the JSON payload. The game
attribute was added to the attributes array to make sure that bets can
be included (the jsonapi-serializer API sees all attributes that have
their own section as relationships [e.g. player as well, although it is
not a ref attribute], and only includes the content if one non
relationship attribute exists). Will file a bug report (if I am not
misunderstanding the JSON API specs).

The team schema was also added to make serializing and including bets
easier. It should be extended in the next steps.