Move Bets serializer opts into a function
The bets need to be serialized in two different places (as a root object
in /api/bets and as a ref object in /api/games). The serializer options
are mostly the same and should be defined in one place (serializer.js).

To achive this, serializerOpts is changed into a function that returns a
different opts object depending if it should be used as a ref or a root
object.

This function is exported and can be imported and used in api/games.

To make it easy for callers, the function returns the ref object when it
is called without a parameter (default behavior).

I used lodash merge to be able to define simple objects for the base
options, the ref options and the root options since they are easier  to
read than lots of baseOptions.attr = ... calls.