Add query middleware for the football-data.org API
API request can be easily handled with the popular 'request' library. I
used 'request-promise' to be able to use await.

The middleware can be configured with the API url (for easier testing
and easier changes if the url ever changes) and an API token. The API
token is loaded from the ENV (like all secrets).

I add the helper function to the `ctx` so all following middleware can
easily access it.