Move SoccerSeasons serializer options out of Games
The games serializer specified how to serialize the referenced
SoccerSeason. This information should be defined in the soccerseasons
serializer. I moved it there and turned it into a function (similar to
bets/serializer).

fp.mergeWith was necessary because the rootOptions should add more
attributes to the attributes array instead of overwriting it.
Idea for concatArray from: https://lodash.com/docs#mergeWith