Create all teams of a soccerseason on create
A soccerseason has teams that participate. To make sure that all
necessary teams are saved in the `teams` collection, they should be
created whenever a new soccerseason is created.

To make it easy to create the teams, the team data from
football-data.org is used (including the ids).

Since teams need to be associated with a soccerseason by adding them to
the teams array, they are added immediately after they are created.