Include Team data in SoccerSeason
When you want information about a SoccerSeason you often also want
information about the teams that participate. That's why the information
should be included.

The Teams serializeOpts had to be changed into a function to use the
rules n the SoccerSeason serializer.

The keyForAttribute fp.identity ensures that all attributes have the
same name as in the database (camelCase as opposed to dash-case, which
is the default serialization of jsonapi-serializer).