Return a populated SoccerSeason after creation
When the SoccerSeason is created with a POST to /api/soccerseasons the
newly created (and returned) resource did not include the teams. To
unify the experience it populates the teams path now.

It is not possible to call populate on the saved document. You have to
do it via Model.populate(...).
See: http://stackoverflow.com/a/17505622