Always include Teams in SoccerSeasons
Teams were not included when a SoccerSeason was referenced before. But
you want to know about the teams of the season most of the time (e.g.
when SoccerSeasons are included in a game it makes sense to also know
about all the available teams, because the players have to choose their
teams).

Thats why the settings about including the teams could move to the
baseOptions.