Populate the SoccerSeason when showing Games
When you list all games it is enough to get basic information about the
related SoccerSeason (maybe to show the name of the season for an easier
overview). That's why I did not include the teams field on the games
list.

However when you just look at one game you propably want to know about
the teams that play this season. That's why they are also populated. You
need a special form to populate multiple levels deep.
See: http://mongoosejs.com/docs/populate.html#deep-populate

I used the options object for the populate call in controller.index
because it is easier to understand in my opinion (only caption league
and year are included).

I removed the toObject() because the result seems to be the same (the
tested game did not include any bets). It might cause troubles with bets
and has to be reverted.