Add global mongoose plugin to query football-data
I want to move requests to the football-data.org API out of the
controllers and into the models. To be able to use the same function on
all models, it is registered as a global plugin.
See: http://mongoosejs.com/docs/plugins.html#global

The plugin makes the footballData available as a static method because
you often want to use it to create documents (so instance methods do not
work). It is possible to use it e.g for a Team model like this:

apiData = await Team.footballData('url')

The plugin needs to be registrered before any Schemas or Models are
created. To achieve this it has to be loaded before the app is loaded.