Move Team creation with API data to Team Model
It makes more sense that the Team model knows how to create documents
when you pass it a soccerseason id. That's why I moved the createTeams
functionality from the soccerseasons controller to the Team model.

I realized that it is not necessary to pass only ids to addToSet so the
createFromSoccerSeason function returns the newly created teams.