Load SoccerSeason API data with middleware
The API call to get the additional SoccerSeason data should also not be
in the controller. That's why they are moved into a middleware function
that loads the data before the document is validated (necessary because
mongoose would complain about missing required fields that are loaded
from the API on validation).

The middleware should use the footballData that is available as a static
method on all Schemas (because of the global plugin) to avoid having to
specify the API key for the API itself (should only be done in one
central location). Middleware has no access to Schema methods. To give
the new middleware access to the Schema method footballData it is added
by a plugin (Plugins have access to the schema).

I planned to fetch the teams from the API as well but middleware has no
access to other models (as far as I know). That's why the SoccerSeason
teams are still fetched by the model function in the SoccerSeasons
controller.