Replace addApiData plugin with middleware
The addApiData plugin was only necessary to have the footballData static
method available. But it is also possible to get the Model of a Document
with doc.constructor so the plugin is no longer necessary.
See: http://stackoverflow.com/a/17685260

The setting of the API data was simplified with the set method of
Document.
See: http://mongoosejs.com/docs/api.html#document_Document-set