Add fixtures to soccerSeasons via middleware
The fixtures should be added to soccerseasons when a find or a findOne
operation is performed. This is done in a post middleware that adds a
field 'fixtures' to the returned documents (disadvantage: clients need
to call document.toObject() before they are serialized to include the
fixtures as well).

Fixtures are not added on .find() when the query is empty (all
soccerSeasons should be shown) because the client probably wants just an
overview und does not need all the information
alreay (/api/soccerseasons/ will just return basic information without
teams or fixtures).