Handle configuration with the 'config' package
Configuration should be in a central location and it should be possible
to have different configurations for the different environments.

The config module allows for different environments (based on NODE_ENV),
makes it possible to use values from the environment (important for the
.env workflow) and allows to use yaml configuration files (are more
readable than json in my opinion).

An alternative could be 'nconf' which allows more fine grained
precedence control (not needed at this point).