Add footballdata plugin on a per model basis
Since it is easy to get access to the configuration now, it makes more
sense to add the footballData plugin only to the models that need it and
pass the configuration multiple times). This makes having a global
plugin unnecessary which makes it easier to understand what each model
can do.