Add a simple /api/fixtures API endpoint
The fixtures endpoint only supports a GET request with an ID at the
moment (e.g. /api/fixtures/:id) because fixtures should not be
created (the are fetched from the API) and it makes no sense for the app
to show all fixtures (of different seasons) because a game is tied to
one soccerseason.

Since fixtures should always be fetched from the API, the model is not a
mongoose model but uses the same methods so the controller code stays
familiar (e.g. findOne).

The fixtures schema also supports the plugin method to be able to use
the same plugins as the mongoose models (especially footballData).

I removed the Id from homeTeam and awayTeam because homeTeam and
awayTeam seem clearer names (when homeTeamName and awayTeamName are
ignored).