Replace fixture team id with simple object
The jsonapi-serializer has trouble to use single values as references
when the parent structure is not also the root structure (homeTeam: 5
works as reference when just one fixture is returned, when the fixture
is resolved as part of e.g. a soccerseason the reference no longer
works). To solve this problem the home- and awayTeam are transformed
into simple objects homeTeam: { id: 5 } and can be referenced even in
nested structures. This avoids having to load the actual team values
from the database even when they should not be used.

The transformation is moved into a transform function to be able to
reuse it (instead of implementing it in findOne like it was before).

The ref values needed to change to reflect that homeTeam and awayTeam
are no longer simple values but objects.