Serialize fixtures correctly with soccerSeasons
Fixtures are added via a post middleware and to a field that is not
included in the Schema. That's why toObject() is necessary (the added
'fixtures' field would not be serialized otherwise).

The typeForAttribute function is necessary because homeTeam and awayTeam
would have their attribute names as type instead of 'teams'.