Avoid creating a bets collection in the database
Bets are included in a game and therefore only need a Schema. To make
them similar to all the other models a Bet model was created before. But
this leads to the creation of a bets collection in the database. Since
this collection is unnecessary it should not be created.